+++ prbt_hardware_support/include/prbt_hardware_support/libmodbus_client.h
@@ -30,50 +30,31 @@ namespace prbt_hardware_support
 /**
  * @brief Wrapper around libmodbus, see https://libmodbus.org/
  *
- * Used by PilzModbusReadClient to access a modbus server
+ * Used by PilzModbusClient to access a modbus server
  */
 class LibModbusClient : public ModbusClient
 {
 public:
 
-  virtual ~LibModbusClient();
+  //! @brief See base class.
+  virtual ~LibModbusClient() override;
 
-  /**
-   * @brief Initialize the modbus client by connecting to the server
-   *
-   * @param ip of the server
-   * @param port to use
-   * @return true if the connection to the server succeeded
-   * @return false if the connection to the server failed
-   */
-  bool init(const char* ip, unsigned int port);
+  //! @brief See base class.
+  bool init(const char* ip, unsigned int port) override;
 
-  /**
-   * @brief Set the response timeout
-   *
-   * Interval to wait for a response from the server if within this timespan no answer is received readHoldingRegisters
-   * will throw a ModbusExceptionDisconnect
-   *
-   * @param timeout_ms in ms
-   */
-  void setResponseTimeoutInMs(unsigned int timeout_ms);
+  //! @brief See base class.
+  void setResponseTimeoutInMs(unsigned long timeout_ms) override;
 
-  /**
-   * @brief Get the response timeout
-   *
-   * @return response timeout
-   */
-  unsigned int getResponseTimeoutInMs();
+  //! @brief See base class.
+  unsigned long getResponseTimeoutInMs() override;
 
-  /**
-   * @brief Read the holding registers
-   *
-   * @param addr starting address to read from
-   * @param nb number of registers to read
-   * @throw ModbusExceptionDisconnect if a disconnect from the server happens
-   * @return std::vector<uint16_t> containing the register contents
-   */
-  std::vector<uint16_t> readHoldingRegister(int addr, int nb);
+  //! @brief See base class.
+  RegCont readHoldingRegister(int addr, int nb) override;
+
+  //! @brief See base class.
+  RegCont writeReadHoldingRegister(const int write_addr,
+                                   const RegCont &write_reg,
+                                   const int read_addr, const int read_nb) override;
 
   /**
    * @brief Close connection with server