+++ prbt_hardware_support/include/prbt_hardware_support/modbus_client.h
@@ -21,12 +21,18 @@
 #include <vector>
 #include <cstdint>
 
+#include "register_container.h"
 
 namespace prbt_hardware_support
 {
 
 class ModbusClient
 {
+public:
+  // LCOV_EXCL_START
+  virtual ~ModbusClient() = default;
+  // LCOV_EXCL_STOP
+
 public:
 
   /**
@@ -47,14 +53,14 @@ public:
    *
    * @param timeout_ms in ms
    */
-  virtual void setResponseTimeoutInMs(unsigned int timeout_ms) = 0;
+  virtual void setResponseTimeoutInMs(unsigned long timeout_ms) = 0;
 
   /**
    * @brief Get the response timeout
    *
    * @return response timeout
    */
-  virtual unsigned int getResponseTimeoutInMs() = 0;
+  virtual unsigned long getResponseTimeoutInMs() = 0;
 
   /**
    * @brief Read the holding registers
@@ -62,9 +68,14 @@ public:
    * @param addr starting address to read from
    * @param nb number of registers to read
    * @throw ModbusExceptionDisconnect if a disconnect from the server happens
-   * @return std::vector<uint16_t> containing the register contents
+   * @returns the registers content.
    */
-  virtual std::vector<uint16_t> readHoldingRegister(int addr, int nb) = 0;
+  virtual RegCont readHoldingRegister(int addr, int nb) = 0;
+
+  virtual RegCont writeReadHoldingRegister(const int write_addr, const RegCont &write_reg,
+                                           const int read_addr, const int read_nb) = 0;
+
+
 
 };