+++ prbt_hardware_support/test/unittests/unittest_modbus_adapter_sto.cpp
@@ -34,9 +34,11 @@
 #include <prbt_hardware_support/param_names.h>
 #include <prbt_hardware_support/ModbusMsgInStamped.h>
 #include <prbt_hardware_support/modbus_msg_sto_wrapper.h>
-#include <prbt_hardware_support/modbus_msg_in_utils.h>
-#include <prbt_hardware_support/sto_modbus_adapter.h>
+#include <prbt_hardware_support/modbus_msg_in_builder.h>
+#include <prbt_hardware_support/modbus_adapter_sto.h>
 #include <prbt_hardware_support/pilz_manipulator_mock.h>
+#include <prbt_hardware_support/register_container.h>
+#include <prbt_hardware_support/wait_for_topic.h>
 
 #include <pilz_testutils/async_test.h>
 
@@ -44,7 +46,7 @@ namespace prbt_hardware_support
 {
 
 static const std::string STO_TOPIC {"/stop1"};
-static const std::string STO_ADAPTER_NODE_NAME {"/sto_modbus_adapter_node"};
+static const std::string STO_ADAPTER_NODE_NAME {"/modbus_adapter_sto_node"};
 
 static constexpr bool STO_CLEAR {true};
 static constexpr bool STO_ACTIVE {false};
@@ -57,78 +59,99 @@ static constexpr int MODBUS_API_VERSION_FOR_TESTING {2};
 using namespace prbt_hardware_support;
 
 using ::testing::_;
+using ::testing::InSequence;
 
 
 #define EXPECT_STOP1 \
+  do { \
   EXPECT_CALL(manipulator_, unholdCb(_,_)).Times(0);\
   EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(0);\
   EXPECT_CALL(manipulator_, holdCb(_,_)).Times(1);\
-  EXPECT_CALL(manipulator_, haltCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("halt_callback"));\
+  EXPECT_CALL(manipulator_, haltCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("halt_callback")); }\
+  while(false)
 
 #define EXPECT_CLEARANCE \
+  do { \
   EXPECT_CALL(manipulator_, holdCb(_,_)).Times(0); \
   EXPECT_CALL(manipulator_, haltCb(_,_)).Times(0); \
   EXPECT_CALL(manipulator_, unholdCb(_,_)).Times(1); \
-  EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("recover_callback")); \
+  EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("recover_callback")); }\
+  while(false)
 
 /**
  * @brief Test fixture class. Sets up a publisher to the modbus_read topic
  * and a manipulator mock which advertises the needed controller and driver services.
  */
-class PilzStoModbusAdapterTest : public ::testing::Test, public ::testing::AsyncTest
+class ModbusAdapterStoTest : public ::testing::Test, public ::testing::AsyncTest
 {
 protected:
-  void SetUp();
+  ModbusAdapterStoTest();
+  virtual ~ModbusAdapterStoTest() override;
+
+  virtual void SetUp() override;
 
   ModbusMsgInStampedPtr createDefaultStoModbusMsg(bool sto_clear);
   std::thread asyncConstructor();
 
 protected:
-  ros::NodeHandle nh_;
+  ros::NodeHandle nh_ {};
   ros::NodeHandle nh_private_ {"~"};
   ros::AsyncSpinner spinner_ {2}; // Keep at 2! Instable if only 1 due the service callbacks in threads
 
-  int index_of_first_register_to_read_;
-  int num_registers_to_read_;
+  int index_of_first_register_to_read_ {0};
+  int num_registers_to_read_ {0};
 
-  ros::Publisher pub_;
+  std::unique_ptr<ModbusAdapterSto> modbus_sto_adapter_ {nullptr};
   ManipulatorMock manipulator_;
+  ros::Publisher pub_ {nh_.advertise<ModbusMsgInStamped>(TOPIC_MODBUS_READ,1)};
 
-  std::string HOLD_SERVICE_T {PilzStoModbusAdapterNode::HOLD_SERVICE};
-  std::string UNHOLD_SERVICE_T {PilzStoModbusAdapterNode::UNHOLD_SERVICE};
-  std::string RECOVER_SERVICE_T {PilzStoModbusAdapterNode::RECOVER_SERVICE};
-  std::string HALT_SERVICE_T {PilzStoModbusAdapterNode::HALT_SERVICE};
+  std::string HOLD_SERVICE_T {ModbusAdapterSto::HOLD_SERVICE};
+  std::string UNHOLD_SERVICE_T {ModbusAdapterSto::UNHOLD_SERVICE};
+  std::string RECOVER_SERVICE_T {ModbusAdapterSto::RECOVER_SERVICE};
+  std::string HALT_SERVICE_T {ModbusAdapterSto::HALT_SERVICE};
 };
 
-void PilzStoModbusAdapterTest::SetUp()
+ModbusAdapterStoTest::ModbusAdapterStoTest()
 {
   spinner_.start();
+}
 
+ModbusAdapterStoTest::~ModbusAdapterStoTest()
+{
+  // Before the destructors of the class members are called, we have
+  // to ensure that all topic and service calls done by the AsyncSpinner
+  // threads are finished. Otherwise, we sporadically will see threading
+  // exceptions like:
+  // "boost::mutex::~mutex(): Assertion `!res' failed".
+  spinner_.stop();
+}
+
+void ModbusAdapterStoTest::SetUp()
+{
   ASSERT_TRUE(nh_private_.getParam(PARAM_INDEX_OF_FIRST_REGISTER_TO_READ_STR, index_of_first_register_to_read_))
       << "No modbus offset given.";
 
   ASSERT_TRUE(nh_private_.getParam(PARAM_NUM_REGISTERS_TO_READ_STR, num_registers_to_read_))
       << "No modbus size given.";
 
-  pub_ = nh_.advertise<ModbusMsgInStamped>(TOPIC_MODBUS_READ,1);
   EXPECT_EQ(0, pub_.getNumSubscribers());
 }
 
-ModbusMsgInStampedPtr PilzStoModbusAdapterTest::createDefaultStoModbusMsg(bool sto)
+ModbusMsgInStampedPtr ModbusAdapterStoTest::createDefaultStoModbusMsg(bool sto)
 {
   static int msg_time_counter {1};
-  std::vector<uint16_t> tab_reg(num_registers_to_read_);
+  RegCont tab_reg(static_cast<uint16_t>(num_registers_to_read_));
   tab_reg[0] = sto;
   tab_reg[1] = MODBUS_API_VERSION_FOR_TESTING;
-  ModbusMsgInStampedPtr msg {createDefaultModbusMsgIn(index_of_first_register_to_read_, tab_reg)};
+  ModbusMsgInStampedPtr msg {ModbusMsgInBuilder::createDefaultModbusMsgIn(static_cast<uint16_t>(index_of_first_register_to_read_), tab_reg)};
   msg->header.stamp = ros::Time(msg_time_counter++);
   return msg;
 }
 
-std::thread PilzStoModbusAdapterTest::asyncConstructor()
+std::thread ModbusAdapterStoTest::asyncConstructor()
 {
   std::thread t([this](){
-    PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+    ModbusAdapterSto adapter_node(nh_, test_api_spec);
   });
   return t;
 }
@@ -137,7 +160,7 @@ std::thread PilzStoModbusAdapterTest::asyncConstructor()
  * @brief Test increases function coverage by ensuring that all Dtor variants
  * are called.
  */
-TEST_F(PilzStoModbusAdapterTest, testModbusMsgWrapperExceptionDtor)
+TEST_F(ModbusAdapterStoTest, testModbusMsgWrapperExceptionDtor)
 {
   std::shared_ptr<ModbusMsgWrapperException> es{new ModbusMsgWrapperException("Test msg")};
 }
@@ -146,20 +169,30 @@ TEST_F(PilzStoModbusAdapterTest, testModbusMsgWrapperExceptionDtor)
  * @brief Test increases function coverage by ensuring that all Dtor variants
  * are called.
  */
-TEST_F(PilzStoModbusAdapterTest, testModbusMsgStoWrapperDtor)
+TEST_F(ModbusAdapterStoTest, testModbusMsgStoWrapperDtor)
 {
   ModbusMsgInStampedConstPtr msg_const_ptr {createDefaultStoModbusMsg(STO_CLEAR)};
   std::shared_ptr<ModbusMsgStoWrapper> ex {new ModbusMsgStoWrapper(msg_const_ptr, test_api_spec)};
 }
 
 /**
+ * @brief Test increases function coverage by ensuring that all Dtor variants
+ * are called.
+ */
+TEST_F(ModbusAdapterStoTest, testAdapterStoDtor)
+{
+  manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
+  std::shared_ptr<AdapterSto> adapter {new AdapterSto(nh_)};
+}
+
+/**
  * @brief Test that the Setup functions properly
  */
-TEST_F(PilzStoModbusAdapterTest, testSetup)
+TEST_F(ModbusAdapterStoTest, testSetup)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
   EXPECT_EQ(1, pub_.getNumSubscribers());
 }
@@ -169,7 +202,7 @@ TEST_F(PilzStoModbusAdapterTest, testSetup)
  *
  * Expected: Constructor blocks until the halt service is available.
  */
-TEST_F(PilzStoModbusAdapterTest, testSetupNoDisableService)
+TEST_F(ModbusAdapterStoTest, testSetupNoDisableService)
 {
   manipulator_.advertiseHoldService(nh_, HOLD_SERVICE_T);
   manipulator_.advertiseUnholdService(nh_, UNHOLD_SERVICE_T);
@@ -188,13 +221,13 @@ TEST_F(PilzStoModbusAdapterTest, testSetupNoDisableService)
  *
  * Expected: Constructor finishes successfully without unhold service.
  */
-TEST_F(PilzStoModbusAdapterTest, testSetupNoUnholdService)
+TEST_F(ModbusAdapterStoTest, testSetupNoUnholdService)
 {
   manipulator_.advertiseHoldService(nh_, HOLD_SERVICE_T);
   manipulator_.advertiseHaltService(nh_, HALT_SERVICE_T);
   manipulator_.advertiseRecoverService(nh_, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
   EXPECT_EQ(1, pub_.getNumSubscribers());
 }
@@ -204,13 +237,13 @@ TEST_F(PilzStoModbusAdapterTest, testSetupNoUnholdService)
  *
  * Expected: Constructor finishes successfully without recover service
  */
-TEST_F(PilzStoModbusAdapterTest, testSetupNoRecoverService)
+TEST_F(ModbusAdapterStoTest, testSetupNoRecoverService)
 {
   manipulator_.advertiseHoldService(nh_, HOLD_SERVICE_T);
   manipulator_.advertiseUnholdService(nh_, UNHOLD_SERVICE_T);
   manipulator_.advertiseHaltService(nh_, HALT_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
   EXPECT_EQ(1, pub_.getNumSubscribers());
 }
@@ -220,7 +253,7 @@ TEST_F(PilzStoModbusAdapterTest, testSetupNoRecoverService)
  *
  * Expected: Constructor blocks until the hold service is available.
  */
-TEST_F(PilzStoModbusAdapterTest, testSetupNoHoldService)
+TEST_F(ModbusAdapterStoTest, testSetupNoHoldService)
 {
   manipulator_.advertiseUnholdService(nh_, UNHOLD_SERVICE_T);
   manipulator_.advertiseHaltService(nh_, HALT_SERVICE_T);
@@ -237,13 +270,13 @@ TEST_F(PilzStoModbusAdapterTest, testSetupNoHoldService)
 /**
  * @brief Tests that a message giving sto clearance is handled correctly
  */
-TEST_F(PilzStoModbusAdapterTest, testClearMsg)
+TEST_F(ModbusAdapterStoTest, testClearMsg)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_CLEARANCE
+  EXPECT_CLEARANCE;
 
   pub_.publish(createDefaultStoModbusMsg(STO_CLEAR));
 
@@ -253,11 +286,11 @@ TEST_F(PilzStoModbusAdapterTest, testClearMsg)
 /**
  * @brief Tests that a message giving sto clearance is handled correctly
  */
-TEST_F(PilzStoModbusAdapterTest, testRemoveUnholdService)
+TEST_F(ModbusAdapterStoTest, testRemoveUnholdService)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
   EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("recover_callback"));
 
@@ -270,11 +303,11 @@ TEST_F(PilzStoModbusAdapterTest, testRemoveUnholdService)
 /**
  * @brief Tests that a message giving sto clearance is handled correctly
  */
-TEST_F(PilzStoModbusAdapterTest, testRemoveRecoverService)
+TEST_F(ModbusAdapterStoTest, testRemoveRecoverService)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
   EXPECT_CALL(manipulator_, unholdCb(_,_)).Times(1).WillOnce(ACTION_OPEN_BARRIER("unhold_callback"));
 
@@ -289,13 +322,13 @@ TEST_F(PilzStoModbusAdapterTest, testRemoveRecoverService)
  *
  * Controller must be holded and driver must be disabled
  */
-TEST_F(PilzStoModbusAdapterTest, testHoldMsg)
+TEST_F(ModbusAdapterStoTest, testHoldMsg)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   pub_.publish(createDefaultStoModbusMsg(STO_ACTIVE));
 
@@ -303,16 +336,16 @@ TEST_F(PilzStoModbusAdapterTest, testHoldMsg)
 }
 
 /**
- * @brief Tests that a message indicating a disconnect from modbus stops the robot even if the content
- * would give sto clearance
+ * @brief Tests that a message indicating a disconnect from modbus stops the
+ * robot even if the content would give sto clearance
  */
-TEST_F(PilzStoModbusAdapterTest, testDisconnectNoStoMsg)
+TEST_F(ModbusAdapterStoTest, testDisconnectNoStoMsg)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_CLEAR);
   msg->disconnect.data = true;
@@ -326,13 +359,13 @@ TEST_F(PilzStoModbusAdapterTest, testDisconnectNoStoMsg)
  * @brief Tests that a message indicating a disconnect from modbus stops the robot when the msg itself
  * would also require sto to go active
  */
-TEST_F(PilzStoModbusAdapterTest, testDisconnectWithStoMsg)
+TEST_F(ModbusAdapterStoTest, testDisconnectWithStoMsg)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_ACTIVE);
   msg->disconnect.data = true;
@@ -346,13 +379,13 @@ TEST_F(PilzStoModbusAdapterTest, testDisconnectWithStoMsg)
  * @brief Tests that a message indicating a disconnect from modbus stops with no
  * other data defined in the message
  */
-TEST_F(PilzStoModbusAdapterTest, testDisconnectPure)
+TEST_F(ModbusAdapterStoTest, testDisconnectPure)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg (new ModbusMsgInStamped());
   msg->header.stamp = ros::Time::now();
@@ -366,13 +399,13 @@ TEST_F(PilzStoModbusAdapterTest, testDisconnectPure)
 /**
  * @brief Tests that stop happens if no version is defined
  */
-TEST_F(PilzStoModbusAdapterTest, testNoVersion)
+TEST_F(ModbusAdapterStoTest, testNoVersion)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_ACTIVE);
   msg->holding_registers.data.pop_back();
@@ -385,13 +418,13 @@ TEST_F(PilzStoModbusAdapterTest, testNoVersion)
 /**
  * @brief Tests that a stop happens if the version is wrong
  */
-TEST_F(PilzStoModbusAdapterTest, testWrongVersion)
+TEST_F(ModbusAdapterStoTest, testWrongVersion)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_ACTIVE);
   msg->holding_registers.data[1] = 0;
@@ -407,13 +440,13 @@ TEST_F(PilzStoModbusAdapterTest, testWrongVersion)
  *
  * @note Version 1 had mistake in specification on the hardware therefore not supported at all
  */
-TEST_F(PilzStoModbusAdapterTest, testVersion1)
+TEST_F(ModbusAdapterStoTest, testVersion1)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_ACTIVE);
   msg->holding_registers.data[1] = 1;
@@ -427,13 +460,13 @@ TEST_F(PilzStoModbusAdapterTest, testVersion1)
 /**
  * @brief Test that stop happends if no Sto is defined
  */
-TEST_F(PilzStoModbusAdapterTest, testNoSto)
+TEST_F(ModbusAdapterStoTest, testNoSto)
 {
   manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
 
-  PilzStoModbusAdapterNode adapter_node(nh_, test_api_spec);
+  modbus_sto_adapter_.reset(new ModbusAdapterSto(nh_, test_api_spec));
 
-  EXPECT_STOP1
+  EXPECT_STOP1;
 
   ModbusMsgInStampedPtr msg = createDefaultStoModbusMsg(STO_ACTIVE);
   msg->holding_registers.data.erase(msg->holding_registers.data.begin());
@@ -445,9 +478,54 @@ TEST_F(PilzStoModbusAdapterTest, testNoSto)
 }
 
 /**
+ * @brief Helper class which allows us to test a function which is normally
+ * protected and, therefore, not directly accessible for testing.
+ */
+class TestingAdapterSto : public AdapterSto
+{
+public:
+  TestingAdapterSto(ros::NodeHandle& nh)
+    : AdapterSto(nh)
+  {}
+
+  virtual ~TestingAdapterSto() = default;
+
+public:
+  void testingUpdateSto(const bool sto)
+  {
+    updateSto(sto);
+  }
+};
+
+/**
+ * @brief Tests that unhold, recover, halt, hold services are ONLY called
+ * if the STO changes its value.
+ */
+TEST_F(ModbusAdapterStoTest, testSameStoValue)
+{
+  manipulator_.advertiseServices(nh_, HOLD_SERVICE_T, UNHOLD_SERVICE_T, HALT_SERVICE_T, RECOVER_SERVICE_T);
+
+  TestingAdapterSto sto_adapter(nh_);
+  {
+    InSequence dummy;
+    EXPECT_CALL(manipulator_, unholdCb(_,_)).Times(1);
+    EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(1);
+    EXPECT_CALL(manipulator_, holdCb(_,_)).Times(0);
+    EXPECT_CALL(manipulator_, haltCb(_,_)).Times(0);
+  }
+  sto_adapter.testingUpdateSto(true);
+
+  EXPECT_CALL(manipulator_, unholdCb(_,_)).Times(0);
+  EXPECT_CALL(manipulator_, recoverCb(_,_)).Times(0);
+  EXPECT_CALL(manipulator_, holdCb(_,_)).Times(0);
+  EXPECT_CALL(manipulator_, haltCb(_,_)).Times(0);
+  sto_adapter.testingUpdateSto(true);
+}
+
+/**
  * @brief Check construction of the exception (essentially for full function coverage)
  */
-TEST_F(PilzStoModbusAdapterTest, ModbusMsgExceptionCTOR)
+TEST_F(ModbusAdapterStoTest, ModbusMsgExceptionCTOR)
 {
   ModbusMsgStoWrapperException* exception = new ModbusMsgStoWrapperException("test");
 
@@ -458,7 +536,7 @@ TEST_F(PilzStoModbusAdapterTest, ModbusMsgExceptionCTOR)
 
 int main(int argc, char** argv)
 {
-  ros::init(argc, argv, "unittest_sto_modbus_adapter");
+  ros::init(argc, argv, "unittest_modbus_adapter_sto");
   ros::NodeHandle nh_;
 
   testing::InitGoogleTest(&argc, argv);